requirements.txt
streamlit
flask
torch
torchvision
Pillow
numpy
opencv-python
reportlab
